hi_tornado/day04/requirement.txt
pymysql
sqlalchemy
tornado==4.5
baidu-aip
